Changed data type to float32, adapted the LeNet encoder, and updated gitignore.

+++ Code/Extractors/Extractor.py
@@ -5,8 +5,8 @@ class Extractor:
         self.processor = processor
         self.dataset_size = dataset_size
         self.feature_shape = feature_shape
-        self.data = numpy.zeros((dataset_size, ) + feature_shape)
-        self.labels = numpy.zeros(dataset_size)
+        self.data = numpy.zeros((dataset_size, ) + feature_shape, dtype=numpy.float32)
+        self.labels = numpy.zeros(dataset_size, dtype=numpy.float32)
         self.class_dict = dict()
 
     def extract(self, dataset_root):
